Extract RunCrateDBQuery function
